Fixed bugs with routes and transaction page
