+++ client/loadData.js
@@ -1,17 +1,16 @@
 import Web3 from 'web3'
 import Fortmatic from 'fortmatic'
+//import axios from 'axios';
+
 // import EtherExchange from "../abis/EtherExchange.json";
 
 let fm = new Fortmatic('pk_test_E28EBDED6FA415DC', 'ropsten')
 
 export const loadBlockchainData = async data => {
-  // hard-coded for testing purposes
-  data.recipient = '0xD3833FA45edd662de6F828BF332DAf4b78Af0365'
   if (window.ethereum) {
     // New Metamask (WebPage)
     window.web3 = new Web3(window.ethereum)
     await window.ethereum.enable()
-
     const web3 = window.web3
     let account = web3.eth.accounts[0]
     console.log('METAMASK WEBPAGE: ', account)
@@ -20,8 +19,17 @@ export const loadBlockchainData = async data => {
     // New Metamask (Extension)
     console.log('METAMASK EXTENSION: ', data.account)
     console.log('RECIPIENT: ', data.recipient)
-
-    return {account: data.account, recipient: data.recipient}
+    //also set user wallet to userdb entry
+    //let recipient = axios.get("/users", data.recipient);
+    if (data.recipient)
+      return {
+        account: data.account,
+        recipient: '0xD3833FA45edd662de6F828BF332DAf4b78Af0365' //recipient wallet
+      }
+    else
+      return {
+        account: data.account
+      }
   } else if (window.web3 || (data && data.web3 !== 'undefined')) {
     // Old Metamask (Both)
     console.log('Please update your Metamask!')
@@ -29,22 +37,29 @@ export const loadBlockchainData = async data => {
   } else if (data) {
     // No Metamask - Fortmatic (Extension)
     window.web3 = new Web3(fm.getProvider())
-
-    window.web3.eth.getAccounts((error, accounts) => {
+    let account = ''
+    let recipient = '0xD3833FA45edd662de6F828BF332DAf4b78Af0365' //hardcoded for tests
+    await window.web3.eth.getAccounts((error, accounts) => {
       if (error) throw error
       console.log('FORTMATIC EXTENSION: ', accounts[0])
       console.log('RECIPIENT: ', data.recipient)
-      //get infortmation from routes, create an axios.get recipient wallet
-      //also set user wallet to userdb entry
-      return {
-        account: accounts[0],
-        recipient: data.recipient //recipient wallet
-      }
+      //set user wallet to userdb entry
+      //let recipient = axios.get("/users", data.recipient);
+      account = accounts[0]
     })
+    if (data.recipient)
+      return {
+        account,
+        recipient //recipient wallet
+      }
+    else
+      return {
+        account
+      }
   } else {
     // No Metamask - Fortmatic (WebPage)
     window.web3 = new Web3(fm.getProvider())
-    window.web3.eth.getAccounts((error, accounts) => {
+    return await window.web3.eth.getAccounts((error, accounts) => {
       if (error) throw error
       console.log('FORTMATIC WEBPAGE: ', accounts[0])
       return {
